backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
psycopg2-binary==2.9.9
Pillow==10.2.0
django-extensions==3.2.3
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
geopy==2.4.1
folium==0.15.1
joblib==1.3.2